Add a code reviewer option into the README
I've done the example in Python and it works.

ChatGPT response: "Sure, I'd be happy to help! In order to give the most useful feedback, it would be helpful if you could provide some context about what the code is supposed to do and any specific areas you would like me to focus on. Is there anything else you can tell me about the code you want me to review?"