visual and navigation enhancements
Hi, I would like to contribute in enhancing this repo and I would be happy to be a part of this amazing space.  please review my commits I created a visual enhancement with nav to reach the contents faster as it grows large in time. it's required to have this nav for easy content reach and accessibility.   thank you, Faraday