Update README.md
Following up on my last enhancement commits, this is a test to take readme navigation into consideration as this repo is blowing up and it’s hard for users to navigate thru contents. But navigation will help with visual and accessibility.  I’d be happy to stay consistent with all the nav, prompts updates and changes. Thanks!