Update README.md with Gretel's Synthetic Multilingual Prompts Dataset
Inspired by the valuable work in this repository and dataset, Gretel has created a synthetic multilingual dataset of prompts. This dataset, released under the permissive Apache 2.0 license, includes 1,250 synthetic LLM prompts generated using Gretel Navigator, available in seven different languages. It is designed to be used with LLMs to generate diverse and multilingual responses based on the provided prompts. We believe this contribution will further enhance and support the efforts made here by providing additional resources for the community to explore, utilize, and build upon.